Added convergence handler
The convergence handler stops training when the error on the training set converges i.e. when the difference between the last iteration and the n previous ones is less than a specified tolerance. Right now the patience is w.r.t. batches. It may be necessary to switch to epochs.